#1524 Make a history util for loading local storage
